Updated both files
In support, I added a function for reading in the yyjj file. Similarly, the code also generates labels for each event, however the logic used was incorrect, leading to problems. This will be changed soon.
In FindTheHiggs, I added all components for the XGBoost model. It currently achieves 100% accuracy due to the incorrect labelling of events, this too will be fixed soon.